+++ src/tables/mod.rs
@@ -23,6 +23,7 @@ pub mod hmtx;
 pub mod kern;
 pub mod loca;
 pub mod maxp;
+pub mod mvar;
 pub mod name;
 pub mod os2;
 pub mod post;